requirements.txt
audiocraft==0.0.2a2
langchain==0.0.214
poe_api==0.4.6
python-dotenv==1.0.0
streamlit==1.23.1
transformers==4.30.2
